src/services/Backend APIS/ocr-service/requirements.txt
spacy==3.5.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
Flask==2.2.5
flask-cors==3.0.10
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3
python-dotenv==1.0.0
opencv-python==4.8.0.76
numpy==1.26.4
